+++ R/GOCore.R
@@ -172,6 +172,55 @@ chord_dat <- function(data, genes, process){
 }
 
 #' 
+#' @name reduce_overlap
+#' @title Eliminates redundant terms.
+#' @description The function eliminates all terms with a gene overlap >= set
+#'   threshold (\code{overlap}) The reduced dataset can be used to improve the
+#'   readability of plots such as \code{GOBubble} and \code{GOBar}
+#' @param data A data frame created with \code{circle_dat}.
+#' @param overlap Skalar indicating the threshold for gene overlap (default = 0.75).
+#' @details The function is currently very slow.
+#' @examples
+#' \dontrun{
+#' # Load the included dataset
+#' data(EC)
+#' 
+#' # Building the circ object
+#' circ <- circle_dat(EC$david, EC$genelist)
+#' 
+#' # Eliminate redundant terms
+#' reduced_circ <- reduce_overlap(circ)
+#' 
+#' # Plot reduced data
+#' GOBubble(reduced_circ)
+#' 
+#' }
+#' @export
+
+reduce_overlap <- function(data, overlap){
+  term <- genes <- NULL
+  if (missing(overlap)) overlap <- 0.75
+  terms <- unique(data$term)
+  FUN <- function(x,y) round(sum(x$genes %in% y$genes)/nrow(x), digits = 2)
+  tmp <- matrix(0, ncol = length(terms), nrow = length(terms), dimnames = list(terms, terms))
+  for (row in 1:nrow(tmp)){
+    for (col in 1:ncol(tmp)){
+      tmp[row, col] <- FUN(subset(data, term == terms[row], genes), subset(data, term == terms[col], genes))
+    }
+  }
+  tmp[base::upper.tri(tmp)] <- 0
+  for(col in 1:ncol(tmp)){
+    idx <- which(tmp[,col] >= overlap)
+    sel_col <- idx[which(idx != col)]
+    tmp[,sel_col] <- 0
+  }
+  sel_terms <- colnames(tmp)[colSums(tmp) != 0]
+  dat <- subset(data, term %in% sel_terms)
+  data <- dat[!duplicated(dat$term), ]
+  return(data)
+}
+
+#' 
 #' @name GOBubble
 #' @title Bubble plot.
 #' @description The function creates a bubble plot of the input \code{data}. The
@@ -183,20 +232,22 @@ chord_dat <- function(data, genes, process){
 #'   plot ('single') or a facet plot with panels for each category 
 #'   (default='single')
 #' @param title The title (on top) of the plot
-#' @param color A character vector which defines the color of the bubbles for 
+#' @param colour A character vector which defines the colour of the bubbles for 
 #'   each category
 #' @param labels Sets a threshold for the displayed labels. The threshold refers
 #'   to the -log(adjusted p-value) (default=5)
 #' @param ID If TRUE then labels are IDs else terms
 #' @param table.legend Defines whether a table of GO ID and GO term should be 
 #'   displayed on the right side of the plot or not (default = TRUE)
-#' @param table.col If TRUE then the table entries are colored according to 
+#' @param table.col If TRUE then the table entries are coloured according to 
 #'   their category, if FALSE then entries are black
-#' @details The x- axis of the plot represents the z-score. The negative
+#' @param bg.col Should only be used in case of a facet plot. If TRUE then the
+#'   panel backgrounds are coloured according to the displayed category
+#' @details The x- axis of the plot represents the z-score. The negative 
 #'   logarithm of the adjusted p-value (corresponding to the significance of the
 #'   term) is displayed on the y-axis. The area of the plotted circles is 
 #'   proportional to the number of genes assigned to the term. Each circle is 
-#'   colored according to its category and labeled alternatively with the ID or 
+#'   coloured according to its category and labeled alternatively with the ID or 
 #'   term name.If static is set to FALSE the mouse hover effect will be enabled.
 #' @import ggplot2
 #' @import gridExtra
@@ -207,24 +258,25 @@ chord_dat <- function(data, genes, process){
 #' data(EC)
 #' 
 #' #Building the circ object
-#' circ<-circular_dat(EC$david, EC$genelist)
+#' circ <- circular_dat(EC$david, EC$genelist)
 #' 
-#' #Creating the bubble plot coloring the table entries according to the category
-#' GOBubble(circ, table.col=T)
+#' #Creating the bubble plot colouring the table entries according to the category
+#' GOBubble(circ, table.col = T)
 #' 
 #' #Creating the bubble plot displaying the term instead of the ID and without the table
-#' GOBubble(circ,ID=F,table.legend=F)
+#' GOBubble(circ, ID = F, table.legend = F)
 #' 
 #' #Faceting the plot
-#' GOBubble(circ, display='multiple')
+#' GOBubble(circ, display = 'multiple')
 #' }
 #' @export
-GOBubble <- function(data, display, title, color, labels, ID = T, table.legend = T, table.col = T){
+GOBubble <- function(data, display, title, colour, labels, ID = T, table.legend = T, table.col = T, bg.col = F){
   zscore <- adj_pval <- category <- count <- id <- term <- NULL
   if (missing(display)) display <- 'single'
   if (missing(title)) title <- ''
-  if (missing(color)) cols <- c("chartreuse4", "brown2", "cornflowerblue") else cols <- color
+  if (missing(colour)) cols <- c("chartreuse4", "brown2", "cornflowerblue") else cols <- colour
   if (missing(labels)) labels <- 5
+  if (bg.col == T & display == 'single') cat("Parameter bg.col will be ignored. To use the parameter change display to 'multiple'")
   
   colnames(data) <- tolower(colnames(data))
   if(!'count'%in%colnames(data)){
@@ -233,40 +285,48 @@ GOBubble <- function(data, display, title, color, labels, ID = T, table.legend =
   }else {rang <- c(1, 30)}
   data$adj_pval <- -log(data$adj_pval, 10)
   sub <- data[!duplicated(data$term), ]
-  g <- ggplot(sub, aes(zscore, adj_pval))+
+  g <- ggplot(sub, aes(zscore, adj_pval, fill = category, size = count))+
     labs(title = title, x = 'z-score', y = '-log (adj p-value)')+
-    geom_point(aes(col = category, size = count), alpha = 1 / 2)+
+    geom_point(shape = 21, col = 'black', alpha = 1 / 2)+
     geom_hline(yintercept = 1.3, col = 'orange')+
     scale_size(range = rang, guide = 'none')
   if (!is.character(labels)) sub2 <- subset(sub, subset = sub$adj_pval >= labels) else sub2 <- subset(sub, sub$id%in%labels | sub$term%in%labels)
   if (display == 'single'){
-    g <- g + scale_colour_manual('Category', values = cols, labels = c('Biological Process', 'Cellular Component', 'Molecular Function'))+
+    g <- g + scale_fill_manual('Category', values = cols, labels = c('Biological Process', 'Cellular Component', 'Molecular Function'))+
       theme(legend.position = 'bottom')+
-      annotate ("text", x = min(sub$zscore), y = 1.35, label = "threshold", colour = "orange", size = 3)
+      annotate ("text", x = min(sub$zscore)+0.2, y = 1.4, label = "Threshold", colour = "orange", size = 4)
     if (ID) g <- g+ geom_text(data = sub2, aes(x = zscore, y = adj_pval, label = id), size = 5) else g <- g + geom_text(data = sub2, aes(x = zscore, y = adj_pval, label = term), size = 4)
     if (table.legend){
       if (table.col) table <- draw_table(sub2, col = cols) else table <- draw_table(sub2)
-      g <- g + theme(axis.text = element_text(size = 14), axis.line = element_line(color = 'grey80'), axis.ticks = element_line(color = 'grey80'), 
+      g <- g + theme(axis.text = element_text(size = 14), axis.line = element_line(colour = 'grey80'), axis.ticks = element_line(colour = 'grey80'), 
                      axis.title = element_text(size = 14, face = 'bold'), panel.background = element_blank(), panel.grid.minor = element_blank(), 
-                     panel.grid.major = element_line(color = 'grey80'), plot.background = element_blank()) 
+                     panel.grid.major = element_line(colour = 'grey80'), plot.background = element_blank()) 
       graphics::par(mar = c(0.1, 0.1, 0.1, 0.1))
       grid.arrange(g, table, ncol = 2)
     }else{
-      g + theme(axis.text = element_text(size = 14), axis.line = element_line(color = 'grey80'), axis.ticks = element_line(color = 'grey80'), 
+      g + theme(axis.text = element_text(size = 14), axis.line = element_line(colour = 'grey80'), axis.ticks = element_line(colour = 'grey80'), 
                 axis.title = element_text(size = 14, face = 'bold'), panel.background = element_blank(), panel.grid.minor = element_blank(), 
-                panel.grid.major = element_line(color = 'grey80'), plot.background = element_blank())
+                panel.grid.major = element_line(colour = 'grey80'), plot.background = element_blank())
     }
   }else{
-    g <- g + facet_grid(.~category, space = 'free_x', scales = 'free_x') + scale_colour_manual(values = cols, guide ='none')
+    if(bg.col){
+      dummy_col <- data.frame(category = c('BP', 'CC', 'MF'), adj_pval = sub$adj_pval[1:3], zscore = sub$zscore[1:3], size = 1:3, count = 1:3)
+      g <- g + geom_rect(data = dummy_col, aes(fill = category), xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf, alpha = 0.1)+
+             facet_grid(.~category, space = 'free_x', scales = 'free_x')+ 
+             scale_fill_manual(values = cols, guide ='none')
+    }else{
+      g <- g + facet_grid(.~category, space = 'free_x', scales = 'free_x')+ 
+               scale_fill_manual(values = cols, guide ='none') 
+    }
     if (ID) {
       g + geom_text(data = sub2, aes(x = zscore, y = adj_pval, label = id), size = 5) + 
-        theme(axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), axis.line = element_line(color = 'grey80'), 
-              axis.ticks = element_line(color = 'grey80'), panel.border = element_rect(fill = 'transparent', color = 'grey80'),
+        theme(axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), axis.line = element_line(colour = 'grey80'), 
+              axis.ticks = element_line(colour = 'grey80'), panel.border = element_rect(fill = 'transparent', colour = 'grey80'),
               panel.background = element_blank(), panel.grid = element_blank(), plot.background = element_blank()) 
     }else{
       g + geom_text(data = sub2, aes(x = zscore, y = adj_pval, label = term), size = 5) + 
-        theme(axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), axis.line = element_line(color = 'grey80'), 
-              axis.ticks = element_line(color = 'grey80'), panel.border = element_rect(fill = 'transparent', color = 'grey80'),
+        theme(axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), axis.line = element_line(colour = 'grey80'), 
+              axis.ticks = element_line(colour = 'grey80'), panel.border = element_rect(fill = 'transparent', colour = 'grey80'),
               panel.background = element_blank(), panel.grid = element_blank(), plot.background = element_blank())
     }
   }
@@ -274,8 +334,8 @@ GOBubble <- function(data, display, title, color, labels, ID = T, table.legend =
 
 #' 
 #' @name GOBar
-#' @title Z-score colored barplot.
-#' @description Z-score colored barplot of terms ordered alternatively by 
+#' @title Z-score coloured barplot.
+#' @description Z-score coloured barplot of terms ordered alternatively by 
 #'   z-score or the negative logarithm of the adjusted p-value
 #' @param data A data frame containing at least the term ID and/or term, the 
 #'   adjusted p-value and the z-score. A possible input can be generated with 
@@ -287,7 +347,7 @@ GOBubble <- function(data, display, title, color, labels, ID = T, table.legend =
 #'   ordered according to the z-scores of the processes. Otherwise the bars are 
 #'   ordered by the negative logarithm of the adjusted p-value
 #' @param title The title of the plot
-#' @param zsc.col Character vector to define the color scale for the z-score of 
+#' @param zsc.col Character vector to define the colour scale for the z-score of 
 #'   the form c(high, midpoint,low)
 #' @details If \code{display} is used to facet the plot the width of the panels 
 #'   will be proportional to the length of the x scale.
@@ -318,13 +378,13 @@ GOBar <- function(data, display, order.by.zscore = T, title, zsc.col){
   colnames(data) <- tolower(colnames(data))
   data$adj_pval <- -log(data$adj_pval, 10)
   sub <- data[!duplicated(data$term), ]
-  
+
   if (order.by.zscore == T) {
     sub <- sub[order(sub$zscore, decreasing = T), ]
     leg <- theme(legend.position = 'bottom')
     g <-  ggplot(sub, aes(x = factor(id, levels = stats::reorder(id, adj_pval)), y = adj_pval, fill = zscore)) +
-      geom_bar(stat = 'identity', color = 'black') +
-      scale_fill_gradient2('z-score', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colorbar(title.position = "top", title.hjust = 0.5), 
+      geom_bar(stat = 'identity', colour = 'black') +
+      scale_fill_gradient2('z-score', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colourbar(title.position = "top", title.hjust = 0.5), 
                            breaks = c(min(sub$zscore), max(sub$zscore)), labels = c('decreasing', 'increasing')) +
       labs(title = title, x = '', y = '-log (adj p-value)') +
       leg
@@ -333,20 +393,20 @@ GOBar <- function(data, display, order.by.zscore = T, title, zsc.col){
     leg <- theme(legend.justification = c(1, 1), legend.position = c(0.98, 0.995), legend.background = element_rect(fill = 'transparent'),
                  legend.box = 'vertical', legend.direction = 'horizontal')
     g <-  ggplot(sub, aes( x = factor(id, levels = reorder(id, adj_pval)), y = zscore, fill = adj_pval)) +
-      geom_bar(stat = 'identity', color = 'black') +
-      scale_fill_gradient2('Significance', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colorbar(title.position = "top", title.hjust = 0.5), breaks = c(min(sub$adj_pval), max(sub$adj_pval)), labels = c('low', 'high')) +
+      geom_bar(stat = 'identity', colour = 'black') +
+      scale_fill_gradient2('Significance', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colourbar(title.position = "top", title.hjust = 0.5), breaks = c(min(sub$adj_pval), max(sub$adj_pval)), labels = c('low', 'high')) +
       labs(title = title, x = '', y = 'z-score') +
       leg
   }
   if (display == 'single'){
-    g + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1), axis.line = element_line(color = 'grey80'), axis.ticks = element_line(color = 'grey80'),
+    g + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1), axis.line = element_line(colour = 'grey80'), axis.ticks = element_line(colour = 'grey80'),
               axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), panel.background = element_blank(), 
               panel.border = element_blank(), panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.background = element_blank())        
   }else{
-    g + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1), axis.line = element_line(color = 'grey80'), axis.ticks = element_line(color = 'grey80'),
-              axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), panel.background = element_blank(), 
-              panel.border = element_blank(), panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.background = element_blank())+
-      facet_grid(.~category, space = 'free_x', scales = 'free_x')
+    g + facet_grid(.~category, space = 'free_x', scales = 'free_x')+
+        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1), axis.line = element_line(colour = 'grey80'), axis.ticks = element_line(colour = 'grey80'),
+            axis.title = element_text(size = 14, face = 'bold'), axis.text = element_text(size = 14), panel.background = element_blank(), 
+            panel.border = element_blank(), panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.background = element_blank())
   }
 }
 
@@ -368,14 +428,14 @@ GOBar <- function(data, display, order.by.zscore = T, title, zsc.col){
 #' @param rad1 The radius of the inner circle (default=2)
 #' @param rad2 The radius of the outer circle (default=3)
 #' @param table.legend Shall a table be displayd or not? (default=TRUE)
-#' @param zsc.col Character vector to define the color scale for the z-score of 
+#' @param zsc.col Character vector to define the colour scale for the z-score of 
 #'   the form c(high, midpoint,low)
-#' @param lfc.col A character vector specifying the color for up- and 
+#' @param lfc.col A character vector specifying the colour for up- and 
 #'   down-regulated genes
 #' @param label.size Size of the segment labels (default=5)
 #' @param label.fontface Font style of the segment labels (default='bold')
 #' @details The outer circle shows a scatter plot for each term of the logFC of 
-#'   the assigned genes. The colors can be changed with the argument 
+#'   the assigned genes. The colours can be changed with the argument 
 #'   \code{lfc.col}.
 #'   
 #'   The \code{nsub} argument needs a bit more explanation to be used wisely. First of 
@@ -399,10 +459,10 @@ GOBar <- function(data, display, order.by.zscore = T, title, zsc.col){
 #' # Creating the circular plot
 #' GOCircle(circ)
 #' 
-#' # Creating the circular plot with a different color scale for the logFC
+#' # Creating the circular plot with a different colour scale for the logFC
 #' GOCircle(circ, lfc.col = c('purple', 'orange'))
 #' 
-#' # Creating the circular plot with a different color scale for the z-score
+#' # Creating the circular plot with a different colour scale for the z-score
 #' GOCircle(circ, zsc.col = c('yellow', 'black', 'cyan'))
 #' 
 #' # Creating the circular plot with different font style
@@ -466,14 +526,14 @@ GOCircle <- function(data, title, nsub, rad1, rad2, table.legend = T, zsc.col, l
     ylim(1, rad2 + 1.6) +
     xlim(0, 10) +
     theme_blank +
-    scale_fill_gradient2('z-score', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colorbar(title.position = "top", title.hjust = 0.5), breaks = c(min(df$zscore), max(df$zscore)),labels = c('decreasing', 'increasing')) +
+    scale_fill_gradient2('z-score', space = 'Lab', low = zsc.col[3], mid = zsc.col[2], high = zsc.col[1], guide = guide_colourbar(title.position = "top", title.hjust = 0.5), breaks = c(min(df$zscore), max(df$zscore)),labels = c('decreasing', 'increasing')) +
     theme(legend.position = 'bottom', legend.background = element_rect(fill = 'transparent'), legend.box = 'horizontal', legend.direction = 'horizontal') +	
-    geom_point(data = dfp, aes(x = logx, y = logy2 + logy), pch = 21, fill = 'transparent', color = 'black', size = 3)+
-    geom_point(data = dfp, aes(x = logx, y = logy2 + logy, color = logFC), size = 2.5)+
+    geom_point(data = dfp, aes(x = logx, y = logy2 + logy), pch = 21, fill = 'transparent', colour = 'black', size = 3)+
+    geom_point(data = dfp, aes(x = logx, y = logy2 + logy, colour = logFC), size = 2.5)+
     scale_colour_manual(values = lfc.col, guide = guide_legend(title.position = "top", title.hjust = 0.5))		
   
   if (table.legend){
-    table <- draw_table(suby, col = 'black')
+    table <- draw_table(suby)
     graphics::par(mar = c(0.1, 0.1, 0.1, 0.1))
     grid.arrange(c, table, ncol = 2)
   }else{